fix: coinbase pay flow
