+++ src/pages/Beta/Buy/CoinbasePay/index.tsx
@@ -1,11 +1,15 @@
-import React, { useEffect, useState } from 'react'
-import { initOnRamp } from '@coinbase/cbpay-js'
+import React, { useEffect, useRef, useState } from 'react'
+import { CBPayInstanceType, initOnRamp } from '@coinbase/cbpay-js'
 import { COINBASE_PK } from 'src/constants'
 import { useActiveWeb3React } from 'src/hooks'
+import { ToggleWalletButton, WalletIcon } from '../styled'
+import { Box, useWalletModalToggle } from '@pangolindex/components'
 
 export default function CoinbasePay() {
   const [init, setInit] = useState(false)
   const { account } = useActiveWeb3React()
+  const coinbasePayInstance = useRef<CBPayInstanceType>()
+  const toggleWalletModal = useWalletModalToggle()
 
   useEffect(() => {
     if (init) return
@@ -30,11 +34,53 @@ export default function CoinbasePay() {
         if (err) {
           console.error(err)
         } else if (instance) {
+          coinbasePayInstance.current = instance
           instance.open()
         }
       }
     )
   }, [account, init])
 
-  return <div />
+  const openCoinbasePay = () => {
+    if (coinbasePayInstance.current) {
+      coinbasePayInstance.current.open()
+    }
+  }
+
+  return (
+    <Box
+      display="flex"
+      flexDirection="column"
+      alignItems="center"
+      maxWidth={350}
+      margin="0px auto"
+      justifyContent="center"
+      width="100%"
+      height="100%"
+      flex={1}
+    >
+      {/* coinbase pay needs an account, so if user is not connected then ask to connect wallet */}
+      {!account ? (
+        <Box>
+          <Box bgColor="black" borderRadius="5px" mb="10px" color="white" p="10px">
+            To use Coinbase Pay first you need to connect your wallet to Pangolin dApp. Please proceed by clicking
+            button below.
+          </Box>
+          <ToggleWalletButton variant="primary" onClick={toggleWalletModal} width="100%">
+            <WalletIcon color="black" />
+            Connect Wallet
+          </ToggleWalletButton>
+        </Box>
+      ) : (
+        <Box>
+          <Box bgColor="black" borderRadius="5px" mb="10px" color="white" p="10px">
+            You will be redirected to Coinbase Pay shortly, if not please click below button
+          </Box>
+          <ToggleWalletButton variant="primary" onClick={openCoinbasePay} width="100%">
+            Open Coinbase Pay
+          </ToggleWalletButton>
+        </Box>
+      )}
+    </Box>
+  )
 }
